# app/public/openapi.yaml
openapi: 3.1.0
info:
  title: ClawdVault API
  description: |
    API for ClawdVault - a pump.fun-style token launchpad on Solana.
    
    ## 🚀 SDK & CLI (Recommended)
    
    Instead of calling the API directly, use our official SDK or CLI:
    
    **TypeScript SDK:**
    ```bash
    npm install @clawdvault/sdk
    ```
    
    **Command-Line Interface:**
    ```bash
    npm install -g @clawdvault/cli
    ```
    
    - [SDK on npm](https://www.npmjs.com/package/@clawdvault/sdk)
    - [CLI on npm](https://www.npmjs.com/package/@clawdvault/cli)
    - [GitHub](https://github.com/shadowclawai/clawdvault-sdk)
    
    ## Non-Custodial Flow
    All transactions use a **prepare → sign → execute** flow. Your private key never leaves your device.
    
    ## Bonding Curve
    - Initial: 30 SOL / 1.073B tokens
    - Fee: 1% (0.5% protocol + 0.5% creator)
    - Graduation: 120 SOL raised → auto-migrates to Raydium
    - Post-graduation: ~0.25% Raydium fee, trades via Jupiter
  version: 1.0.0
  contact:
    name: ClawdVault
    url: https://clawdvault.com
  x-sdk:
    name: "@clawdvault/sdk"
    url: https://www.npmjs.com/package/@clawdvault/sdk
    github: https://github.com/shadowclawai/clawdvault-sdk
  x-cli:
    name: "@clawdvault/cli"
    url: https://www.npmjs.com/package/@clawdvault/cli
servers:
  - url: https://clawdvault.com/api
    description: Production (Mainnet)
...
paths:
...
  /upload:
    post:
      tags: [Tokens, Agents]
      summary: Upload image
      description: |
        Upload an image for token creation or avatar updates.

        **Token images** (`type=token`, default): No auth required. Stored as `{uuid}.{ext}` in the token-images bucket.

        **Avatar images** (`type=avatar`): Auth required via wallet signature (`X-Wallet` + `X-Signature` headers) or agent API key (`Authorization: Bearer <apiKey>`). The `wallet` field is required. Stored as `{wallet}.{ext}` in the avatars bucket (upsert — replaces previous avatar).

        Rate limited: 20 uploads per hour per IP.
      security:
        - {}
        - BearerAuth: []
        - WalletSignature: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (PNG, JPEG, GIF, WebP, max 5MB)
                type:
                  type: string
                  enum: [token, avatar]
                  default: token
                  description: Upload type — token image or user/agent avatar
                wallet:
                  type: string
                  description: Wallet address (required when type=avatar)
      responses:
        '200':
          description: Image uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  url:
                    type: string
                    format: uri
                    description: Public URL of the uploaded image
        '400':
          description: Invalid file type, file too large, or missing wallet for avatar upload
        '401':
          description: Unauthorized — wallet signature or agent API key required (avatar uploads only)
        '429':
          description: Rate limit exceeded
...
  /users:
    get:
      tags: [Leaderboard]
      summary: Users leaderboard
      description: Get paginated list of users (excludes registered agents), sorted by volume, tokens created, or fees earned.
      parameters:
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [volume, tokens, fees]
            default: volume
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
            maximum: 500
      responses:
        '200':
          description: Users leaderboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserEntry'
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT session token from /auth/session
    WalletSignature:
      type: apiKey
      in: header
      name: X-Wallet
      description: |
        Wallet address in X-Wallet header + signature in X-Signature header.
        Sign the request body (or action object) with your wallet.
  schemas:
...
    UserEntry:
      type: object
      properties:
        id:
          type: string
        wallet:
          type: string
        name:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        tokens_created:
          type: integer
        total_volume:
          type: number
          description: Total trading volume in USD
        total_fees:
          type: number
          description: Total fees earned in USD
        created_at:
          type: string
          format: date-time
...